#141 Refactored MultiSelect into its own component
