+++ UInnovateApp/src/components/settingsPage/UserMangementTab.tsx
@@ -1,7 +1,5 @@
 import { Tab, Tabs } from "react-bootstrap";
 import { Button, FormControl, Select, Switch, MenuItem, Chip, Stack } from "@mui/material"
-import CancelIcon from "@mui/icons-material/Cancel";
-import CheckIcon from "@mui/icons-material/Check";
 import TableComponent from "react-bootstrap/Table";
 import "../../styles/TableComponent.css";
 import "../../styles/UserManagementTab.css";
@@ -13,6 +11,7 @@ import { useSelector } from "react-redux";
 import { RootState } from "../../redux/Store";
 import { Role } from "../../redux/AuthSlice";
 import UnauthorizedScreen from "../UnauthorizedScreen";
+import MultiSelect from "./MultiSelect";
 
 
 
@@ -130,55 +129,7 @@ const UserTableRow: React.FC<UserTableRowProps> = ({ firstName, lastName, emailA
 	</tr >
 }
 
-interface MultiSelectProps {
-	selectedList: string[],
-	setSelectedList: React.Dispatch<React.SetStateAction<string[]>>,
-	choiceList: string[],
-	size?: "small" | "medium"
-}
 
-// Component obtained from https://codesandbox.io/s/mui-multi-select-kptq04?from-embed=&file=%2Fsrc%2FApp.js%3A238-291
-export const MultiSelect: React.FC<MultiSelectProps> = ({ selectedList, setSelectedList, choiceList, size }) => {
-
-	return (
-			<Select
-				size={size}
-				multiple
-				value={selectedList}
-				onChange={(e) => setSelectedList((e.target.value as string[]).sort((a, b) => a.localeCompare(b)))}
-				inputProps={{ 'aria-label': 'Without label' }}
-				sx={({ 
-					 '& .MuiSelect-select': { }, minWidth:'20vw', maxWidth: '25vw' })}
-				renderValue={(selected) => (
-					<Stack gap={1} direction="row" flexWrap="wrap">
-						{selected.map((value) => (
-							<Chip
-								key={value}
-								label={value}
-								onDelete={() =>
-									setSelectedList(
-										selectedList.filter((item) => item !== value)
-									)
-								}
-								deleteIcon={
-									<CancelIcon
-										onMouseDown={(event) => event.stopPropagation()}
-									/>
-								}
-							/>
-						))}
-					</Stack>
-				)}
-			>
-				{choiceList.map((choice) => (
-					<MenuItem key={choice} value={choice} sx={{ justifyContent: "space-between" }}>
-						{choice}
-						{selectedList.includes(choice) ? <CheckIcon color="info" /> : null}
-					</MenuItem>
-				))}
-			</Select>
-	);
-}
 
 
 export default UserManagementTab;
